+++ Pandora/Paradigm/Structure/Ability/Morphable.hs
@@ -1,17 +1,17 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
 
-module Pandora.Paradigm.Structure.Ability.Convertible where
+module Pandora.Paradigm.Structure.Ability.Morphable where
 
 import Pandora.Core.Functor (type (~>))
 import Pandora.Pattern.Category ((.))
 import Pandora.Paradigm.Primary.Functor.Tagged (Tagged (Tag))
 import Pandora.Paradigm.Schemes.TU (TU (TU), type (<:.>))
 
-class Convertible f t where
-	type Conversion (f :: k) (t :: * -> *) :: * -> *
-	conversion :: Tagged f <:.> t ~> Conversion f t
+class Morphable f t where
+	type Morphing (f :: k) (t :: * -> *) :: * -> *
+	morphing :: Tagged f <:.> t ~> Morphing f t
 
-convert :: forall f t . Convertible f t => t ~> Conversion f t
-convert = conversion . TU . Tag @f
+morph :: forall f t . Morphable f t => t ~> Morphing f t
+morph = morphing . TU . Tag @f
 
 data Walk a = Preorder a | Inorder a | Postorder a | Levelorder a